Arreglado conceptos de cliente y foto

+++ src/main/java/backend/JDA/repositorios/FotoRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import backend.JDA.modelo.Foto;
 
 @Repository
-public interface FotoRepositorio extends CrudRepository<Foto, Long> {
+public interface FotoRepositorio extends CrudRepository<Foto, Integer> {
 
 }
